backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
httpx>=0.27.0
google-adk>=0.2.0
langchain>=1.0.2
langchain-community>=0.4.0
langchain-google-genai>=2.0.0
google-cloud-aiplatform>=1.70.0
google-cloud-firestore>=2.21.0
google-cloud-storage>=2.18.0
google-cloud-vision>=3.7.0
spacy>=3.7.0
medspacy>=1.0.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=3.0.0
pypdf>=4.3.0
python-multipart>=0.0.9
pillow>=10.4.0
structlog>=24.4.0
prometheus-client>=0.20.0
